requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Data science (uncomment as needed)
# pandas>=2.0.0
# numpy>=1.24.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

# Web development (uncomment as needed)
# flask>=3.0.0
# fastapi>=0.100.0
# uvicorn>=0.23.0
# requests>=2.31.0

# AI/ML (uncomment as needed)
# openai>=1.0.0
# anthropic>=0.18.0
# langchain>=0.1.0
# transformers>=4.35.0

# Database (uncomment as needed)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# pymongo>=4.5.0

# Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# pylint>=3.0.0

# Utilities
# pydantic>=2.0.0
# click>=8.1.0
# rich>=13.0.0